+++ src/renderer/Renderer.hpp
@@ -12,7 +12,7 @@ class Renderer {
 public:
     void setBackgroundColor(const float red, const float green, const float blue, const float alpha);
     void clear() const;
-    void draw(ChunkManager &chunkManager, Shader& shader, float deltaTime) const;
+    void draw(ChunkManager &chunkManager, Shader& shader, double deltaTime) const;
 
 private:
     void setUniforms(const std::unique_ptr<Object> &object, Shader& shader) const;
